[refactor] fetchOtpSignUp関数をViewから分離し、apiディレクトリ内に移動
